Rename SYCL category files
