+++ core/unit_test/sycl/TestSYCLHostUSM_SharedAlloc.cpp
@@ -42,5 +42,5 @@
 //@HEADER
 */
 
-#include <TestSYCLHostUSMSpace_Category.hpp>
+#include <TestSYCLHostUSM_Category.hpp>
 #include <TestSharedAlloc.hpp>
